Added missing filter match index to properly reset depth measurement and calibrate gyro.

+++ inc/rov2016_Accelerometer.h
@@ -47,5 +47,5 @@ extern void magnetometer_init(void);
 extern void accelerometer_updateValue(void);
 extern void magnetometer_updateValue(void);
 extern int16_t accelerometer_getRawData(uint8_t axis);
-extern float magnetometer_getData(uint8_t axis);
+extern int16_t magnetometer_getData(uint8_t axis);
 extern uint8_t lsm303dlhc_getValues(void);
